Implementation of unlimited syntax

+++ lextypes.go
@@ -41,8 +41,15 @@ type OrganizationAddress struct {
 }
 
 type UserLicense struct {
-    AllowedActivations int   `json:"allowedActivations"`
-    AllowedDeactivations int `json:"allowedDeactivations"`
-    Key string               `json:"key"`
-    Type string              `json:"type"`
+    // The allowed activations of the license. A value of -1 indicates unlimited number of activations.
+    AllowedActivations int64   `json:"allowedActivations"`
+
+    // The allowed activations of the license. A value of -1 indicates unlimited number of deactivations.
+    AllowedDeactivations int64 `json:"allowedDeactivations"`
+
+    // The license key.
+    Key string                 `json:"key"`
+
+    // The license type (node-locked or hosted-floating).
+    Type string                `json:"type"`
 }
